Move assistant API to a separate file.
