+++ lib/services/local/users.go
@@ -37,7 +37,6 @@ import (
 	"golang.org/x/crypto/bcrypt"
 	"golang.org/x/crypto/ssh"
 
-	"github.com/gravitational/teleport/api/client/proto"
 	"github.com/gravitational/teleport/api/types"
 	wantypes "github.com/gravitational/teleport/api/types/webauthn"
 	"github.com/gravitational/teleport/api/utils/keys"
@@ -1379,178 +1378,6 @@ func (s *IdentityService) CreateUserRecoveryAttempt(ctx context.Context, user st
 		Key:     backend.Key(webPrefix, usersPrefix, user, recoveryAttemptsPrefix, uuid.New().String()),
 		Value:   value,
 		Expires: attempt.Expires,
-	}
-
-	_, err = s.Create(ctx, item)
-	return trace.Wrap(err)
-}
-
-// Conversation is a conversation entry in the backend.
-type Conversation struct {
-	Title          string    `json:"title,omitempty"`
-	ConversationID string    `json:"conversation_id"`
-	CreatedTime    time.Time `json:"created_time"`
-}
-
-// CreateAssistantConversation creates a new conversation entry in the backend.
-func (s *IdentityService) CreateAssistantConversation(ctx context.Context, username string,
-	req *proto.CreateAssistantConversationRequest,
-) (*proto.CreateAssistantConversationResponse, error) {
-	if username == "" {
-		return nil, trace.BadParameter("missing parameter username")
-	}
-
-	conversationID := uuid.New().String()
-	payload := &Conversation{
-		ConversationID: conversationID,
-		CreatedTime:    req.CreatedTime,
-	}
-
-	value, err := json.Marshal(payload)
-	if err != nil {
-		return nil, trace.Wrap(err)
-	}
-
-	item := backend.Item{
-		Key:   backend.Key(assistantConversationPrefix, username, conversationID),
-		Value: value,
-	}
-
-	_, err = s.Create(ctx, item)
-	if err != nil {
-		return nil, trace.Wrap(err)
-	}
-
-	return &proto.CreateAssistantConversationResponse{Id: conversationID}, nil
-}
-
-func (s *IdentityService) getConversation(ctx context.Context, username, conversationID string) (*Conversation, error) {
-	item, err := s.Get(ctx, backend.Key(assistantConversationPrefix, username, conversationID))
-	if err != nil {
-		return nil, trace.Wrap(err)
-	}
-
-	var conversation Conversation
-	if err := json.Unmarshal(item.Value, &conversation); err != nil {
-		return nil, trace.Wrap(err)
-	}
-
-	return &conversation, nil
-}
-
-// SetAssistantConversationTitle sets the given title as the assistant conversation title.
-func (s *IdentityService) SetAssistantConversationTitle(ctx context.Context, username string, request *proto.ConversationInfo) error {
-	if request.Id == "" || request.Title == "" {
-		return trace.BadParameter("missing conversation ID or title")
-	}
-
-	msg, err := s.getConversation(ctx, username, request.Id)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	// Only update the title, leave the rest of the fields intact.
-	msg.Title = request.Title
-
-	payload, err := json.Marshal(msg)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	item := backend.Item{
-		Key:   backend.Key(assistantConversationPrefix, username, request.Id),
-		Value: payload,
-	}
-
-	if _, err = s.Update(ctx, item); err != nil {
-		return trace.Wrap(err)
-	}
-
-	return nil
-}
-
-// GetAssistantConversations returns all conversations started by a user.
-func (s *IdentityService) GetAssistantConversations(ctx context.Context, username string, _ *proto.GetAssistantConversationsRequest) (*proto.GetAssistantConversationsResponse, error) {
-	startKey := backend.Key(assistantConversationPrefix, username)
-	result, err := s.GetRange(ctx, startKey, backend.RangeEnd(startKey), backend.NoLimit)
-	if err != nil {
-		return nil, trace.Wrap(err)
-	}
-
-	conversationsIDs := make([]*proto.ConversationInfo, 0, len(result.Items))
-	for _, item := range result.Items {
-		payload := &Conversation{}
-		if err := json.Unmarshal(item.Value, payload); err != nil {
-			return nil, trace.Wrap(err)
-		}
-
-		conversationsIDs = append(conversationsIDs, &proto.ConversationInfo{
-			Id:          payload.ConversationID,
-			Title:       payload.Title,
-			CreatedTime: payload.CreatedTime,
-		})
-	}
-
-	sort.Slice(conversationsIDs, func(i, j int) bool {
-		return conversationsIDs[i].CreatedTime.Before(conversationsIDs[j].GetCreatedTime())
-	})
-
-	return &proto.GetAssistantConversationsResponse{
-		Conversations: conversationsIDs,
-	}, nil
-}
-
-// GetAssistantMessages returns all messages with given conversation ID.
-func (s *IdentityService) GetAssistantMessages(ctx context.Context, username, conversationId string) (*proto.GetAssistantMessagesResponse, error) {
-	if username == "" {
-		return nil, trace.BadParameter("missing username")
-	}
-
-	if conversationId == "" {
-		return nil, trace.BadParameter("missing conversation ID")
-	}
-
-	startKey := backend.Key(assistantMessagePrefix, username, conversationId)
-	result, err := s.GetRange(ctx, startKey, backend.RangeEnd(startKey), backend.NoLimit)
-	if err != nil {
-		return nil, trace.Wrap(err)
-	}
-
-	out := make([]*proto.AssistantMessage, len(result.Items))
-	for i, item := range result.Items {
-		var a proto.AssistantMessage
-		if err := json.Unmarshal(item.Value, &a); err != nil {
-			return nil, trace.Wrap(err)
-		}
-		out[i] = &a
-	}
-
-	sort.Slice(out, func(i, j int) bool {
-		// Sort by created time.
-		return out[i].CreatedTime.Before(out[j].GetCreatedTime())
-	})
-
-	return &proto.GetAssistantMessagesResponse{
-		Messages: out,
-	}, nil
-}
-
-// CreateAssistantMessage adds the message to the backend.
-func (s *IdentityService) CreateAssistantMessage(ctx context.Context, username string, msg *proto.AssistantMessage) error {
-	if username == "" {
-		return trace.BadParameter("missing username")
-	}
-
-	value, err := json.Marshal(msg)
-	if err != nil {
-		return trace.Wrap(err)
-	}
-
-	messageID := uuid.New().String()
-
-	item := backend.Item{
-		Key:   backend.Key(assistantMessagePrefix, username, msg.ConversationId, messageID),
-		Value: value,
 	}
 
 	_, err = s.Create(ctx, item)
