+++ lib/services/local/users_test.go
@@ -28,7 +28,6 @@ import (
 	"github.com/go-webauthn/webauthn/protocol"
 	"github.com/google/go-cmp/cmp"
 	"github.com/google/uuid"
-	"github.com/gravitational/teleport/api/client/proto"
 	"github.com/gravitational/trace"
 	"github.com/jonboulle/clockwork"
 	"github.com/stretchr/testify/require"
@@ -938,83 +937,4 @@ func TestIdentityService_GetKeyAttestationDataV11Fingerprint(t *testing.T) {
 	retrievedAttestationData, err := identity.GetKeyAttestationData(ctx, key.Public())
 	require.NoError(t, err)
 	require.Equal(t, attestationData, retrievedAttestationData)
-}
-
-// TestAssistantCRUD tests the assistant CRUD operations.
-func TestAssistantCRUD(t *testing.T) {
-	t.Parallel()
-
-	identity := newIdentityService(t, clockwork.NewFakeClock())
-	ctx := context.Background()
-
-	const username = "foo"
-	var conversationID string
-
-	t.Run("create conversation", func(t *testing.T) {
-		req := &proto.CreateAssistantConversationRequest{
-			CreatedTime: time.Now(),
-		}
-
-		conversationResp, err := identity.CreateAssistantConversation(ctx, username, req)
-		require.NoError(t, err)
-		require.NotEmpty(t, conversationResp.Id)
-
-		conversationID = conversationResp.Id
-	})
-
-	t.Run("get conversation", func(t *testing.T) {
-		conversations, err := identity.GetAssistantConversations(ctx, username, nil)
-		require.NoError(t, err)
-		require.Len(t, conversations.Conversations, 1)
-		require.Equal(t, conversationID, conversations.Conversations[0].Id)
-	})
-
-	t.Run("create message", func(t *testing.T) {
-		msg := &proto.AssistantMessage{
-			CreatedTime:    time.Now(),
-			ConversationId: conversationID,
-			Payload:        "foo",
-			Type:           "USER_MSG",
-		}
-		err := identity.CreateAssistantMessage(ctx, username, msg)
-		require.NoError(t, err)
-	})
-
-	t.Run("get messages", func(t *testing.T) {
-		messages, err := identity.GetAssistantMessages(ctx, username, conversationID)
-		require.NoError(t, err)
-		require.Len(t, messages.Messages, 1)
-		require.Equal(t, "foo", messages.Messages[0].Payload)
-	})
-
-	t.Run("set conversation title", func(t *testing.T) {
-		titleReq := &proto.ConversationInfo{
-			Title: "bar",
-			Id:    conversationID,
-		}
-		title := "bar"
-		err := identity.SetAssistantConversationTitle(ctx, username, titleReq)
-		require.NoError(t, err)
-
-		conversations, err := identity.GetAssistantConversations(ctx, username, nil)
-		require.NoError(t, err)
-		require.Len(t, conversations.Conversations, 1)
-		require.Equal(t, title, conversations.Conversations[0].Title)
-	})
-
-	t.Run("conversations are sorted by created_time", func(t *testing.T) {
-		req := &proto.CreateAssistantConversationRequest{
-			CreatedTime: time.Now().Add(time.Hour),
-		}
-
-		conversationResp, err := identity.CreateAssistantConversation(ctx, username, req)
-		require.NoError(t, err)
-		require.NotEmpty(t, conversationResp.Id)
-
-		conversations, err := identity.GetAssistantConversations(ctx, username, nil)
-		require.NoError(t, err)
-		require.Len(t, conversations.Conversations, 2)
-		require.Equal(t, conversationID, conversations.Conversations[0].Id)
-		require.Equal(t, conversationResp.Id, conversations.Conversations[1].Id)
-	})
 }
